@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>net5.0</TargetFramework>
+        <TargetFramework>net6.0</TargetFramework>
         <Authors>Chien Zhang</Authors>
         <RepositoryUrl>https://github.com/ac682/Ac682.Extensions.Logging.Console</RepositoryUrl>
         <Title>Ac682.Extensions.Logging.Console</Title>
@@ -12,12 +12,12 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="GitVersion.MsBuild" Version="5.6.10">
+      <PackageReference Include="GitVersion.MsBuild" Version="5.8.1">
         <PrivateAssets>all</PrivateAssets>
         <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
       </PackageReference>
-      <PackageReference Include="Microsoft.Extensions.Logging" Version="5.0.0" />
-      <PackageReference Include="Spectre.Console" Version="0.41.0" />
+      <PackageReference Include="Microsoft.Extensions.Logging" Version="6.0.0" />
+      <PackageReference Include="Spectre.Console" Version="0.43.0" />
     </ItemGroup>
 
 </Project>
