@@ -8,11 +8,12 @@ namespace Ac682.Extensions.Logging.Console.Tests
         private static void Main()
         {
             var builder = new ConsoleLoggerOptionsBuilder();
+            builder.SetMinimalLevel(LogLevel.Information);
             builder.AddBuiltinFormatters();
             var provider = new ConsoleLoggerProvider(builder.Build());
 
             var logger = provider.CreateLogger(typeof(Program).FullName);
-            logger.LogInformation("[blue] I AM BLUE, YOUR ARE {}[/]","[yellow]NOT[/]");
+            logger.LogTrace("[blue] I AM BLUE, YOUR ARE {}[/]","[yellow]NOT[/]");
             logger.LogDebug("Hello, World!");
             logger.LogInformation("I am {Master}({Version}), your {Name}", 123, typeof(Program).Assembly.GetName().Version, "master");
             logger.LogWarning("Now is {NowDateTime}. And I am {Color}", DateTime.Now, ConsoleColor.Blue);
